Disable auto-update
An version of the auto-update was based on a weekly checkout of the master branch.
We wanted to update the master branch as soon as nobody was using this old version any more, as we do not support it any more for security purpose, but it appears that at least one person uses it, so we gently disable the auto-update before starting working on the master branch again (which is useful to show regular evolution of our codebase on the GitHub front page).

Please wait one week after this commit before modifying the master branch and merging to it the last updates so that the old versions have checked this commit out and then deactivated the auto-update.